+++ source/redux/dev tools.js
@@ -1,13 +1,11 @@
+import React from 'react'
 import { createDevTools } from 'redux-devtools'
 import LogMonitor from 'redux-devtools-log-monitor'
 import DockMonitor from 'redux-devtools-dock-monitor'
 
-export default function dev_tools()
-{
-	return createDevTools
-	(
-		<DockMonitor toggleVisibilityKey="ctrl-H" changePositionKey="ctrl-Q">
-			<LogMonitor />
-		</DockMonitor>
-	)
-}
+export default createDevTools
+(
+	<DockMonitor toggleVisibilityKey="ctrl-H" changePositionKey="ctrl-Q" defaultIsVisible>
+		<LogMonitor theme="tomorrow" />
+	</DockMonitor>
+)